backend/requirements.txt
black==19.10b0
dj-rest-auth==1.0.7
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-extensions==2.2.9
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ptipython==1.0.1
pylint==2.5.3
pylint-django==2.0.15
Werkzeug==1.0.1
